Segment/Threat Hunting/Hunt Any Threat/requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
loguru==0.7.3
requests==2.32.5
urllib3==2.5.0
python-dotenv>=1.0.0
pytest>=8.0.0
